DN-4: Simplified and modularised auth views components, can now use the AuthViews component separately from Modal.
