+++ client/src/components/ui/modal/auth/AuthViews/views.tsx
@@ -5,7 +5,7 @@ import {
     Input,
     SubmitButton,
     SwitchViewButton
-} from '../forms';
+} from '../../forms';
 
 interface AuthViewProps {
     formData: any;
